Separate reading out into a separate service
